refactor: split visual tests into separate files by category
Replace the monolithic visual.test.ts with five focused test files:
- elements.test.ts — HTML elements, fragments, styles
- components.test.ts — counter, stateless, composition, memoization
- context.test.ts — Provider, useContext selector/transform
- ui-patch.test.ts — global and local patch lifecycle
- use-effect.test.ts — useEffect AbortSignal tests

Extract shared esbuild bundle setup into fixtures.ts using Playwright
custom fixtures.

Closes #86
